SIG-1503 initial refactor
